requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart
moviepy>=1.0.3
pillow>=9.0.0
python-docx>=0.8.11